Fixes prov #18 EnDat Motor leads to HW error
A hardware status number can be found in the SystemDump.xml file which is not documented anywhere. Currently the number 8 was found with ZF modules and also with EnDat motors. Most likely this means that this hardware is not configured identically in the Automation Studio project, but still everything is fine! -> No error. So as a result everything seems to be displayed correctly now.